[REFACTOR] : 저장한 아티클 어댑터 vm 의존성 제거후 클릭 이벤트 추가  #124
