+++ app/src/main/java/com/indipage/presentation/savedspace/SavedSpaceAdapter.kt
@@ -5,14 +5,11 @@ import android.view.ViewGroup
 import androidx.recyclerview.widget.ListAdapter
 import androidx.recyclerview.widget.RecyclerView.ViewHolder
 import com.example.core_ui.view.ItemDiffCallback
-import com.indipage.data.dto.response.ResponseSpaceDto
-import com.indipage.data.dto.response.SavedSpace
 import com.indipage.databinding.ItemSavedSpaceBinding
 import com.indipage.domain.entity.Space
 
 class SavedSpaceAdapter(
-//    private val viewModel:SavedSpaceViewModel,
-    private val onMoveToDetailClick: (Space,Int) -> Unit = { _,_ -> }
+    private val onMoveToSpaceDetailClick: (Space, Int) -> Unit = { _, _ -> }
     ):
     ListAdapter<Space, SavedSpaceAdapter.SavedSpaceViewHolder>(
     SavedSpaceDiffCalback
@@ -33,9 +30,8 @@ class SavedSpaceAdapter(
     ) : ViewHolder(binding.root) {
         fun bind(data: Space) {
             binding.savedSpace = data
-//            binding.vm= viewModel
             binding.root.setOnClickListener {
-                onMoveToDetailClick(data,position)
+                onMoveToSpaceDetailClick(data,position)
             }
             binding.executePendingBindings()
         }
